fix: Salary Slip w/pay per leave not correctly calculating leave without pay
